start refactoring model fitting code
