+++ 09_simulations/models.py
@@ -2,16 +2,9 @@ import numpy as np
 import pandas as pd
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.linear_model import LogisticRegression
-from sklearn.metrics import (
-    roc_auc_score,
-    average_precision_score,
-    accuracy_score,
-    balanced_accuracy_score
-)
-from sklearn.model_selection import KFold, GridSearchCV, RandomizedSearchCV
+from sklearn.model_selection import GridSearchCV, RandomizedSearchCV
 from sklearn.neural_network import MLPClassifier
 from sklearn.pipeline import Pipeline
-from tqdm.auto import tqdm
 
 def train_ridge(X_train,
                 y_train,
@@ -279,143 +272,4 @@ def train_linear_csd(X_train,
 
     return cv_pipeline
 
-
-def get_prob_metrics(y_train, y_test, y_pred_train, y_pred_test):
-    """Get metrics for predicted probabilities.
-
-    y_pred_train and y_pred_test should be continuous; true values are binary.
-    """
-    train_auroc = roc_auc_score(y_train, y_pred_train, average="weighted")
-    test_auroc = roc_auc_score(y_test, y_pred_test, average="weighted")
-
-    train_aupr = average_precision_score(y_train, y_pred_train, average="weighted")
-    test_aupr = average_precision_score(y_test, y_pred_test, average="weighted")
-
-    return {
-        'train_auroc': train_auroc,
-        'test_auroc': test_auroc,
-        'train_aupr': train_aupr,
-        'test_aupr': test_aupr,
-    }
-
-
-def train_k_folds_all_models(xs, ys, domains, train_data=None, n_splits=4, seed=42):
-    """Split data into k folds and evaluate all implemented models.
-
-    Arguments:
-    xs is a numpy array of features (n, p)
-    ys is a column vector of labels (n, 1)
-    domains is a column vector of domains (n, 1)
-    train_data has the format (xs_train, ys_train, domains_train)
-    and overrides train split if provided
-    """
-    results = []
-    results_cols = None
-    kf = KFold(n_splits=n_splits, shuffle=True, random_state=seed)
-
-    outer_progress = tqdm(enumerate(kf.split(xs)), total=n_splits)
-    inner_progress = tqdm(total=4, leave=True)
-    for fold, (train_ix, test_ix) in outer_progress:
-        outer_progress.set_description('fold: {}'.format(fold))
-        # if train_data is provided then just split test set,
-        # and use provided training data
-        # otherwise split for both train and test
-        if train_data is not None:
-            # NOTE when train_data is passed in we really only have
-            # to train the models once, then evaluate them on each test
-            # fold - could speed things up to move this outside the loop
-            # for long-training models/large datasets
-            X_train, X_test = train_data[0], xs[test_ix, :]
-            y_train, y_test = train_data[1], ys[test_ix, :]
-            ds_train, ds_test = train_data[2], domains[test_ix, :]
-        else:
-            X_train, X_test = xs[train_ix, :], xs[test_ix, :]
-            y_train, y_test = ys[train_ix, :], ys[test_ix, :]
-            ds_train, ds_test = domains[train_ix, :], domains[test_ix, :]
-
-        inner_progress.reset()
-        inner_progress.set_description('model: ridge')
-
-        # train/evaluate ridge regression model
-        fit_pipeline = train_ridge(X_train, y_train.flatten(), seed=seed)
-        y_pred_train = fit_pipeline.predict(X_train)
-        y_pred_test = fit_pipeline.predict(X_test)
-        metrics = get_prob_metrics(y_train, y_test, y_pred_train, y_pred_test)
-
-        metric_cols = list(metrics.keys()) + ['model', 'fold']
-        metric_vals = list(metrics.values()) + ['ridge', fold]
-        if results_cols is None:
-            results_cols = metric_cols
-        else:
-            assert metric_cols == results_cols
-        results.append(metric_vals)
-        inner_progress.update(1)
-        inner_progress.set_description('model: random forest')
-
-        # train/evaluate random forest model
-        fit_pipeline = train_rf(X_train, y_train.flatten(), seed=seed)
-        y_pred_train = fit_pipeline.predict(X_train)
-        y_pred_test = fit_pipeline.predict(X_test)
-        metrics = get_prob_metrics(y_train, y_test, y_pred_train, y_pred_test)
-
-        metric_vals = list(metrics.values()) + ['random_forest', fold]
-        if results_cols is None:
-            results_cols = metric_cols
-        else:
-            assert metric_cols == results_cols
-        results.append(metric_vals)
-        inner_progress.update(1)
-        inner_progress.set_description('model: mlp')
-
-        # train/evaluate 3-layer NN model
-        fit_pipeline = train_mlp(X_train,
-                                 y_train.flatten(),
-                                 seed=seed,
-                                 n_folds=-1,
-                                 max_iter=100)
-        y_pred_train = fit_pipeline.predict_proba(X_train.astype(np.float32))[:, 1]
-        y_pred_test = fit_pipeline.predict_proba(X_test.astype(np.float32))[:, 1]
-        metrics = get_prob_metrics(y_train, y_test, y_pred_train, y_pred_test)
-
-        metric_vals = list(metrics.values()) + ['mlp', fold]
-        if results_cols is None:
-            results_cols = metric_cols
-        else:
-            assert metric_cols == results_cols
-        results.append(metric_vals)
-        inner_progress.update(1)
-        inner_progress.set_description('model: linear CSD')
-
-        # train/evaluate linear model with CSD loss layer
-        fit_pipeline = train_linear_csd(X_train,
-                                        y_train.flatten(),
-                                        ds_train.flatten(),
-                                        seed=seed,
-                                        n_folds=-1,
-                                        max_iter=100)
-        # predict_proba expects the first column of the feature matrix to be the
-        # domain info, so we'll concatenate it here
-        y_pred_train = fit_pipeline.predict_proba(
-            np.concatenate((ds_train, X_train), 1).astype(np.float32)
-        )[:, 1]
-        y_pred_test = fit_pipeline.predict_proba(
-            np.concatenate(
-                (np.zeros((X_test.shape[0], 1)), X_test), 1
-            ).astype(np.float32)
-        )[:, 1]
-        metrics = get_prob_metrics(y_train, y_test, y_pred_train, y_pred_test)
-
-        metric_vals = list(metrics.values()) + ['linear_csd', fold]
-        if results_cols is None:
-            results_cols = metric_cols
-        else:
-            assert metric_cols == results_cols
-        results.append(metric_vals)
-        inner_progress.update(1)
-
-    inner_progress.close()
-    results_df = pd.DataFrame(results, columns=results_cols)
-    results_df = results_df.melt(id_vars=['model', 'fold'], var_name='metric')
-
-    return results_df
